+++ Kernel/Output/HTML/TicketOverviewSmall.pm
@@ -1310,12 +1310,24 @@ sub Run {
             );
         }
 
-        # customer info (customer name)
+        # customer info
+        my %CustomerInfo;
         if ( $Param{Config}->{CustomerInfo} ) {
             if ( $Article{CustomerUserID} ) {
                 $Article{CustomerName} = $Self->{CustomerUserObject}->CustomerName(
                     UserLogin => $Article{CustomerUserID},
                 );
+
+                %CustomerInfo = $Self->{CustomerUserObject}->CustomerUserDataGet(
+                    User => $Article{CustomerUserID},
+                );
+
+                INFOKEY:
+                for my $InfoKey ( sort keys %CustomerInfo ) {
+                    next INFOKEY if $InfoKey =~ m{\ACustomer}xms;
+
+                    $CustomerInfo{ 'Customer' . $InfoKey } = $CustomerInfo{$InfoKey};
+                }
             }
         }
 
@@ -1498,7 +1510,9 @@ sub Run {
                         . $ResponsibleInfo{'UserLastname'};
                 }
                 else {
-                    $DataValue = $Article{$TicketColumn} || $UserInfo{$TicketColumn};
+                    $DataValue = $Article{$TicketColumn}
+                        || $UserInfo{$TicketColumn}
+                        || $CustomerInfo{$TicketColumn};
                 }
 
                 $Self->{LayoutObject}->Block(